refactor(brokers): hoist the rolling-window rate limiter into one module
Phase 2 of the tech-debt plan. Flattrade and Dhan each carried a 90-line
_RollingWindowRateLimiter; there is now one in
Dependencies/broker_rate_limit.py.

THE "VERBATIM COPY" CLAIM WAS WRONG, AND CHECKING MATTERED

The audit called the two copies verbatim. They are not: the docstrings
differ in three places. What IS equal is the part that matters -- parsed
with docstrings and comments stripped and the broker name normalised, the
two bodies are identical. That is the check this hoist rests on, and it
is why this is a de-duplication rather than a behaviour change. Dhan's
wording was kept because it was already broker-neutral ("because both
limits apply at once" rather than naming the broker).

The class reads no module-level constant -- verified by walking its AST
for free names -- so it moved without dragging any policy with it. Only
the budgets differ, and they were already constructor arguments:
Flattrade 40/200 API and 10/40 order, Dhan 20/800 and 10/250.

THE COVERAGE TRAP WAS REAL, AND THEN IT WAS NOT

Moving ~45 well-tested statements out of an adapter leaves the harder-to-
test broker-specific remainder behind, so per-adapter coverage can fall
below its 80% floor even as the total improves. Measured rather than
guessed, on the same percent_covered field the policy gate reads:

  Flattrade  82.3% -> 81.1%   (-1.2)
  Dhan       82.5% -> 81.5%   (-1.1)
  broker_rate_limit.py  97.0% (new, given its own 90% row)

I nearly mis-called this. The pyproject note that "a local run reads ~2
points high" would have put both under 80% on CI. But that offset comes
from 7 SDK-gated tests (neo_api_client, pyotp) that skip on CI, and those
are KOTAK and SHOONYA tests. Pulling the numbers from the last CI run on
this branch confirms it: CI measured Flattrade 82.3% and Dhan 82.5%,
identical to local, while Kotak differs by 1.0pp. So the offset does not
apply to these two files and the post-hoist figures stand as measured.
`check_coverage_thresholds.py` passes locally against the real report.

The new module carries its own 90% row rather than riding on the
adapters' 80%: it sits in front of every live broker request those two
make, so under-counting breaches the broker's published limits and
over-waiting makes a trading decision stale. The dict's own header warns
that a module can escape its budget by moving files -- this is that case.

It also repeats broker_contract's dual sys.modules registration, so the
diagnose_*_symbol.py scripts, which run adapters standalone off a
sys.path bootstrap, cannot end up with a second copy of the class.

Both adapters keep `_RollingWindowRateLimiter = RollingWindowRateLimiter`
under the historical private name: three suites construct it as
`<module>._RollingWindowRateLimiter`, including
Tests/test_nifty_multi_strategy_master.py:8342. `deque` became unused in
both and was removed; `Callable` is still needed in both.

Net -139 lines across the two adapters, +90 shared. Gates: 547 master,
28 market-data-health, 1252 pytest, coverage policy, ruff, mypy,
compileall, bandit all clean.
