+++ Dependencies/Dhan API/dhan_execution.py
@@ -83,7 +83,6 @@ import os
 import sys
 import threading
 import time
-from collections import deque
 from collections.abc import Callable
 from concurrent.futures import Future, ThreadPoolExecutor
 from concurrent.futures import TimeoutError as FuturesTimeoutError
@@ -108,6 +107,14 @@ _REPO_ROOT = _DEPENDENCIES_DIR.parent
 for _import_root in (_REPO_ROOT, _DEPENDENCIES_DIR):
     if str(_import_root) not in sys.path:
         sys.path.insert(0, str(_import_root))
+
+from broker_rate_limit import RollingWindowRateLimiter  # noqa: E402
+
+# Kept under the historical private name: three suites construct it as
+# `<module>._RollingWindowRateLimiter` (Tests/Dependencies/<Broker> API/... and
+# Tests/test_nifty_multi_strategy_master.py), and the adapter's own call sites
+# read better unchanged. The implementation now lives in broker_rate_limit.
+_RollingWindowRateLimiter = RollingWindowRateLimiter
 
 from broker_contract import (  # noqa: E402
     TERMINAL_BROKER_STATES,
@@ -231,98 +238,6 @@ def _env_str(name: str, default: str = "") -> str:
     if len(value) >= 2 and value[0] == value[-1] and value[0] in {"'", '"'}:
         value = value[1:-1].strip()
     return value or default
-
-
-class _RollingWindowRateLimiter:
-    """Enforce per-second and per-minute request budgets across worker threads.
-
-    Short bursts wait for a slot.  A wait longer than ``max_wait_seconds`` raises
-    before the request is sent; a stale live order is more dangerous than a
-    clear indeterminate result that freezes new live entry.
-    """
-
-    def __init__(
-        self,
-        per_second: int,
-        per_minute: int,
-        max_wait_seconds: float,
-        clock: Callable[[], float] = time.monotonic,
-        sleeper: Callable[[float], None] = time.sleep,
-        label: str = "API",
-    ) -> None:
-        self.per_second = int(per_second)
-        self.per_minute = int(per_minute)
-        self.max_wait_seconds = float(max_wait_seconds)
-        self._clock = clock
-        self._sleep = sleeper
-        self.label = label
-        self._timestamps: deque[float] = deque()
-        self._lock = threading.Lock()
-
-    def acquire(self, deadline: float | None = None) -> None:
-        """Reserve one request slot or raise before a long/stale wait.
-
-        The deque stores only timestamps from the last minute.  We separately
-        count its last one-second slice because both limits apply at once.  A
-        short wait is acceptable; a long wait would make a trading decision
-        stale, so it raises *before* any broker request is sent.
-        """
-        started = self._clock()
-        if deadline is None:
-            acquired = self._lock.acquire()
-        else:
-            remaining = deadline - started
-            acquired = remaining > 0 and self._lock.acquire(timeout=remaining)
-        if not acquired:
-            raise TimeoutError(
-                f"Dhan {self.label} deadline expired waiting for the limiter lock"
-            )
-        try:
-            while True:
-                now = self._clock()
-                # Discard calls that are older than the longest (60-second)
-                # window.  Keeping the deque small also makes each check cheap.
-                minute_cutoff = now - 60.0
-                while self._timestamps and self._timestamps[0] <= minute_cutoff:
-                    self._timestamps.popleft()
-
-                # The minute deque includes the one-second window, so select its
-                # recent tail instead of maintaining a second source of truth.
-                recent_second = [
-                    stamp for stamp in self._timestamps if stamp > now - 1.0
-                ]
-                second_full = len(recent_second) >= self.per_second
-                minute_full = len(self._timestamps) >= self.per_minute
-
-                if not second_full and not minute_full:
-                    # Reserving the timestamp before returning prevents two
-                    # strategy threads from claiming the same final slot.
-                    self._timestamps.append(now)
-                    return
-
-                waits = []
-                if second_full:
-                    waits.append(1.0 - (now - recent_second[0]))
-                if minute_full:
-                    waits.append(60.0 - (now - self._timestamps[0]))
-                wait_seconds = max(0.0, max(waits))
-                elapsed = now - started
-                exceeds_call_deadline = (
-                    deadline is not None and now + wait_seconds > deadline
-                )
-                if elapsed + wait_seconds > self.max_wait_seconds or exceeds_call_deadline:
-                    raise RuntimeError(
-                        f"Dhan {self.label} rate limit exhausted; "
-                        f"safe slot needs {wait_seconds:.2f}s"
-                    )
-                self._sleep(wait_seconds)
-        finally:
-            self._lock.release()
-
-    def reset(self) -> None:
-        """Forget local request history after a session is explicitly closed."""
-        with self._lock:
-            self._timestamps.clear()
 
 
 def _classify_envelope(envelope: Any) -> tuple[str, Any, str]:
